COMP: Remove minor typecasting compiler warnings.

+++ Source/MediaStorageAndFileFormat/gdcmBitmap.cxx
@@ -286,7 +286,7 @@ unsigned long Bitmap::GetBufferLength() const
     {
     assert( PF.GetSamplesPerPixel() == 1 );
     const size_t bytesPerRow = Dimensions[0] / 8 + (Dimensions[0] % 8 != 0 ? 1 : 0);
-    const unsigned int save = bytesPerRow * Dimensions[1];
+    const size_t save = bytesPerRow * Dimensions[1];
     if(Dimensions[0] % 8 == 0 )
       assert( save * 8 == mul );
     mul = save;
